Remove extra line
This .good file was written based on scopeResolve generating output per symbol
with the same name.  However, if the naming conflict comes from a use statement
entirely, the line will be the same.  It isn't necessary to point out both
versions of the name in this case, so we don't.