Check dest != src for bulk transfer
Adds a check to confirm that the destination and source pointers are not
identical. If they are the same, do nothing.

Previously we would compare the 'data' or 'shiftedData' pointers,
instead of the offset pointers. This is especially bad for aliases, and
would prevent bulk transfers from non-overlapping slices of the same
array.

Also changes the _ddata_shift function to no longer require an explicit
type argument, and instead relies on a query expression to infer the
element type.