Add an error message future related to naming conflicts from uses
If two modules define a symbol 'foo' and both are used by a single module, and
that module tries to access 'foo' without specifying which 'foo' is meant, this
will correctly cause a naming conflict.  However, the complaint in the current
error message seems to imply that it is wrong for the same name to be used in
multiple modules, which would be confusing to the user.  We would like the
blame to instead be given to the use statements which caused these two symbols
to come into conflict.

Thanks to Vass for pointing out that this was an unclear error message.