Add missing 'input' timer to ISx benchmark
In spite of my concern to get the graphs right for ISx the first
time, I failed to get one of the subtimers for creating the input
due to checking to make sure that the number of keys was equal
to the number of subtimers rather than the number of all timers.
Dur.

I've simultaneously updated the .dat files in the internal chapel-perf
repo to include the new column.