Default to ugni when using the module and CCE is the target compiler
We used to default to gasnet for CCE because we couldn't build ugni. Well,
technically we couldn't build tcmalloc, but ugni required tcmalloc. With the
switch to jemalloc, we can now build ugni support under CCE. We're already
building the module with ugni+cce support, this just changes the default.

We can now use ugni with all the target compilers we currently build the module
with. We could remove the compiler check in chpl_comm.py all together, but I
decided to leave it in on the off chance we ever add pgi support back into the
module build.