Update AstTraverse classes to account for renamed symbols in uses
I forgot to do this when I was adding the renaming capability.  Update is
similar to that of the 'only' and 'except' update - add a helper function for
the two AST traversal options that share code, and make sure that we respond
correctly to whether or not there are other names before the renamed symbols
in the use statement.

Also updated some comments so that the rename data structure is also referred
to, instead of just the normal 'only' or 'except' data structure