Add some tests of a bug found by a user
A user discovered an issue with nested renaming uses where in some cases the
ordering of mutually exclusive uses of the same module would affect whether
certain symbols could be found.  Of the four orderings, only backToSame.chpl
would pass with the compiler as it stood.

Commit with the fix incoming.

(Thanks for finding this, Paul!)