Fix for the renaming error found by our user
It turns out that this error had occurred because I had forgotten to update
UseStmt::providesNewSymbols to take the renaming list into account.  Rectifying
that oversight.

The behavior should be that if one of the uses renames symbols and the other
provides an except list, we should automatically assume we provide new symbols
since even if the new name matches something that was not in the except list,
it is available in a different form than previously.  The same goes for if one
use renames a symbol and the other names the old_name or new_name in their
only list.  The only case where a renaming would potentially disqualify a use
is if the renaming is duplicated in the other use.  Checks for if the count of
duplicates should take the renaming list into account as well.