Add a few assert()s to put Coverity's mind to rest
Coverity flagged five new issues due to two root issues this weekend,
though it seems unlikely to have been due to our code changes.

In one case (qio.c), it seems to be a case that it would
understandably not be able to trace through to understand that the
non-NULL-ness of a pointer is related to whether or not the error
code is zero.

In the other case, it seems surprising that it isn't treating
chpl_internal_error() as an exit() when (it seems to me) it
unambiguously is.

In both cases, I'm guessing/hoping that adding an assertion that a
pointer is non-NULL will quiet it down.