Add a test to ensure that valgrind is working
Add a simple test to verify that valgrind is working by explicitly accessing
memory that hasn't been allocated and expecting an error from valgrind.

We recently discovered that because our test systems don't have the valgrind
header files installed, jemalloc valgrind support wasn't working. As a result,
valgrind couldn't track allocations/deallocations. There was no big indication
that valgrind support was missing, but luckily there was a single future that
was expecting a leak in the .bad file, which allowed us to discovered the
issue.

This test is intended to always run under valgrind so that have some more
assurance that our valgrind results can be trusted.