Call parent_insert_help when adding formals or args
 CallExpr::insertAtHead
 CallExpr::insertAtTail
 FnSymbol::insertFormalAtHead
 FnSymbol::insertFormalAtTail

all add an argument/formal to the AST, but did not immediately update the
newly added AST node to record its position in the tree. In particular,
Expr's parentSymbol and parentExpr fields were not immediately updated.
This change calls parent_insert_help to set these fields as appropriate
immediately. This may not be strictly necessary but can be viewed as an
aid to debugging.