Add four tests of bad behavior with nested limited uses that cover a module
A user discovered an issue with nested mutually exclusive uses of the same
module, where not all of the symbols in the module would be found and which
symbols were found varied depending on the order of these uses.  These tests
exercise all four possible orderings of the use statements.