Fix the nested mutually exclusive use error
This error occurred as a result of adding a use to the alreadySeen map when the
outer use had caused it to not include any symbols at all (for example, when the
outer use excluded the only symbol included by that particular use, and vice
versa), which meant that a later nested look at the same use would fail to be
added to the uses to search for that particular symbol.  I made that choice
while under the impression that failing to add these uses would cause us to
repeatedly encounter them over and over again, greatly slowing down
compilation to the point where it was unreasonable.  Testing does not show that
to be the case now, though.