Halving/Doubling style array vector operations
Implement the array vector operations using a buffer that grows and shrinks
by a factor of its current size instead of one element at a time.  Currently,
the growth factor defaults to 1.5, but it can be changed with a config
param.

Updated timeVectorArray.chpl test to include testing the remove() function.