Convert our extern chpl_mem_* declarations to use size_t instead of int.
The runtime chpl_mem_* functions use size_t, but our externs were declared with
int(64). We still want the chapel interfaces to use int(64), so we safeCast
before we call the extern routines.

Note that we still want the Chapel interface to use ints to avoid the usual
problems associated with working with unsigned types. We're not really losing
anything by using ints, because there's no feasible way you can have an
allocation that's max(int) on 64 bit platforms, much less anything bigger. On
32 bit platforms, size_t is typically only 32 bits, so you could truncate the
int(64) while casting, but you can't actually have an allocation that big on 32
platforms anyways, so it doesn't matter.

This is the second step towards having the extern declarations of the
chpl_mem_* routines match the runtime signatures.