Convert our extern chpl_mem_* declarations to use chpl_mem_descInt_t
The runtime chpl_mem_* functions use chpl_mem_descInt_t, but our externs were
declared with int(16). This wasn't really a problem since chpl_mem_descInt_t is
typedeffed to int(16) in the runtime, but it's nice to be consistent with type
names. This just declares chpl_mem_descInt_t as an entern type and uses that
instead of int(16)

This is the third and last step towards having the extern declarations of the
chpl_mem_* routines match the runtime signatures.