Clean up an old problem in the FileSystem module
I found myself with some spare minutes last night and decided to make sure some
of the FileSystem proposal was up to date.  While verifying that, I discovered
an old note in the copy() implementation about a problem that has since been
resolved.  This commit cleans up that implementation as indicated by the note.

The old problem was a bug about cleaning up channels in functions that can exit
before the channels were created.  Since that bug has been resolved, the
channels no longer needed to be protected by a wrapper scope.