Reduce the number of definitions for chpl_memhook_md_num()
Instead of declaring chpl_memhook_md_num() inside each function that uses it,
just make it a private extern.

Also adds a TODO for chpl_mem_descInt_t to remove the explicit `= int(16)` in
the modules. I tried teasing it away, but it's actually really hard because the
compiler hard-codes it to int(16) as well, which makes not declaring it like
that in the modules tricky. I also tried creating an explicit type in the
compiler to represent it, but that was also annoying because then I needed to
add support for math operations, _defaultOf, and a few other things. I think
what we will probably want to do is define the type in the compiler and just
export it to the runtime, instead of trying to have the compiler figure out
what the runtime is using.