Replace chpl_mem_* with chpl_here_* functions in the string module
The string module externed out to the chpl_mem_* functions to support alloc,
dealloc, etc. This was because the chpl_here_* variants weren't trivially
usable because of type mismatches. Some previous updates have improved the
signature of the chpl_here_* functions and to allow implicit coercions between
c_ptr's and c_void_ptr's. This simply converts all chpl_mem_* calls in the
string module to use the chpl_here_* variants.

This allows some really nice cleanup as it removes the extern declarations and
removes a ton of safeCasts that are now handled in the chpl_here_* functions.

Using chpl_here_* in the string module did require me to explicitly declare the
return types of the chpl_here_* functions to avoid some recursive resolution. I
think I only really needed it for alloc and realloc, but I decided to add it to
all of them to be consistent.