Allow c_ptr's to be implicitly coerced to c_void_ptr
This is to support using the chpl_here_* routines instead of externing to the
chpl_mem_* functions in the string module. Currently the string module defines
the chpl_mem_* routines as taking/return a c_ptr instead of a c_void_ptr,
because we didn't coerce c_ptr's to c_void_ptr's.