Add string wrappers for memmove, memcpy, and memcmp
Previously the string module externed memmove, memcpy, and memcmp and directly
used those functions. This lead to memmove, memcpy, and memcmp being leaked
into the global namespace and led to using a bunch of safeCasts at the
callsites.

This wraps those functions with a simple private inline proc that handles the
safeCasts. This prevents the functions names from escaping into the global
namespace and removes a bunch of safecasts.