Have chpl_str_comm_get take len: int instead of len: size_t
chpl_str_comm_get is a helper routine that the string module uses instead of
directly calling the chpl_comm_get primitive. Instead of having all the call
sites do a safeCast to size_t, just have it take an int and do the safeCast
internally.

This removes the last of the safeCasts from the string code (with the exception
of the helpers that do the casting)