Convert certain reductions to forall loops.
A reduction like this:

      + reduce X;

is converted in the parser to:

      var temp;
      forall x in X with (+ reduce temp) do
        x += temp;
      // reduction result is in temp

with forall intents.

The following reductions are still handled the classic way, i.e. as before:

* over zippered items
* over a forall expression
* with user-defined reduction ops

COMPILER CHANGES

* buildReduceExpr() now starts by calling buildReduceViaForall().
The latter converts the reduction to a forall if it can.
Otherwise the former proceeds as before.

See an AST example below.

* A major implementation challenge was to handle reductions over
promoted expressions. I was forced to deal with them right away :(
because I could not detect promotion in buildReduceViaForall() and
exclude them, as I did other cases. A good deal of the code that
this change adds is due to that.

* Another challenge was that a classic reduction computes the result
- and so resolves the result type - after the loop body.
Whereas a forall with a reduce intent requires that the
reduce-intent variable be declared before the loop body
- and so its type must be declared beforehand too.

The straightforward thing is to make the reduction result type
be the type of the things being reduced. However, that does
not work when +-reducing booleans, e.g. test/trivial/deitz/monte.chpl

My solution results in the AST that does this before the forall loop:

    fn "chpl_reduceNNN"
    {
      // built in buildReduceScanPreface1()
      def chpl_toReduce  // data <-- variable in compiler source
      def chpl_eltType   // eltType
      chpl_toReduce = // dataExpr
      eltType = typeof(... iteratorIndex ... chpl_toReduce)

      // built in buildReduceScanPreface2()
      def chpl_reduceGlob  // globalOp
      chpl_reduceGlob = new REDUCE_SCAN_OP_CLASS(eltType = chpl_eltType)

      // built in buildReduceViaForall() proper
      def chpl_reduceResult : typeof(chpl_reduceGlob.generate()) // result
      def chpl_reduceIndexVar  // index

      // built in buildForallLoopStmt()
      {
        if chpl__tryToken
          standalone case
        else
          if chpl__tryToken
            leader-follower case
          else
            serial case  // added in addElseClauseForSerialIter()
        chpl_reduceResult = chpl_reduceGlob.generate()
      }

      return chpl_reduceResult
    }

To do that, I propagate chpl_globalOp via the useThisGlobalOp argument
all the way into setupOneReduceIntent(), because otherwise
it would create its own chpl_globalOp. Which could work only
when the reduce result has the type of the things being reduced.

* addElseClauseForSerialIter() adds a serial case - to handle
reductions over things that are iterable only serially.
Which is allowed for reductions - given that we have tests
that do that, e.g. test/trivial/deitz/monte.chpl
However, it is not allowed for forall loops, so just
buildForallLoopStmt() wouldn't handle it.
BTW Brad says "maybe it should be allowed?"

* When PRIM_TO_LEADER resolves to a call to _iterator_for_loopexprNN(),
do not do extendLeader() on it. Simply pass the extra args from
origToLeaderCall into the _toLeader call that's within it.
Done in propagateExtraArgsForLoopIter().

BTW _iterator_for_loopexprNN is created in buildLeaderIteratorFn()
invoked from buildForallLoopExpr().

* The change to functionResolution.cpp w.r.t. NamedExpr
handles a case that now happens when actuals are passed
as named expressions.

That NamedExpr was created in preFold() here:

        } else if (call->isPrimitive(PRIM_TO_STANDALONE)) {
          FnSymbol* iterator = getTheIteratorFn(call);
          CallExpr* standaloneCall = new CallExpr(iterator->name);
          for_formals(formal, iterator) {
            standaloneCall->insertAtTail(new NamedExpr(formal->name, new SymExpr(formal)));
          }
        ...

* Future work: I notice that when buildPromotionWrapper() builds
leader and follower iterators, the leader iterator is the result of
_toLeader/Zip() on the arguments to the promoted operation. Instead
the leader should simply be the leader of the first of these arguments.

* Future work: some of the reductions that are still handled the classic
way may be amenable to the new implementation. Try that.

* Future work: replace tryToken with the can-resolve primitive.

TEST CHANGES

* This test now invokes one standalone; used to invoke two followers.
One standalone is correct and better. Adjusted .good correspondingly.

    arrays/deitz/parallelism/test_reduction_is_parallel.good

* Improvements in comm counts. Adjusted .good correspondingly.

    distributions/robust/arithmetic/performance/multilocale/*.good

* Added new tests in:

    test/reductions/vass

The non-future test records today's behavior, which computes
the reductions correctly. However, in a couple of cases
we would like to invoke the standalone iterator for performance,
and that does not happen.

Specifically, in some situations old-style reductions
are still being generated, and they do not look at
standalone iterators. So when the iterator being reduced
has serial + standalone versions, the serial version is invoked.
When it has serial + leader/follower + standlone versions,
the leader/follower version is invoked.

The two futures request that the standalone version be invoked.