Scope resolution changes for ability to use enums
Adds some generality to the code, so that it no longer depends on using a
module.  Renamed some functions to reflect this more general state.

In processImportExprs, I differentiate between using a module and using an enum
when determining whether to add the used symbol to the enclosing ModuleSymbol's
moduleUseList - I decided to do this because that list is for initializing
modules in the proper order and so including enums there wouldn't be necessary.

Fixed the error messages when the symbol being used was unexpected so that they
did not state that only modules can be used.

Changed getUsedModuleSymbol from a function which took a symbol and would
print an error or return you a ModuleSymbol to a function which took a symbol
and would print an error or return true if the symbol could be in a use
statement.

Made getUsedModule able to return either a TypeSymbol which pointed to an
EnumType or a ModuleSymbol, and changed its name to reflect that it wasn't just
going to return a ModuleSymbol.

In UseStmt::validateList, I updated the scope being searched for the symbols
to respond appropriately to whether what is being used is a module or an enum.
This same block of code is later used in lookup, so I should make it a method
on the UseStmt type.

I changed buildBreadthFirstModuleList's alreadySeen map to have a more general
key set (ModuleSymbol or TypeSymbol which is to an EnumType instead of just
ModuleSymbols), and made it vary whether to expect use chains based on if the
use is of a module or an enum.

With this change, the tests which do not hide functions all pass.  Still need
to update function resolution.