Update domainMaps memory tracking test .goods
The array-as-vec work added a new domain member to DefaultRectangular arrays.
This test tracks leaked memory from operations on different domain maps,
and in some cases DefaultRectangular arrays are leaked in it.  Since the
leaked arrays have a new domain field, the amount leaked for each leaked
array went up.