Clean up the duplication I inserted
Adds a method to the UseStmt node to obtain the appropriate scope to search
for symbols.  I had implemented this action in a couple of places in
scopeResolve, but it seemed more appropriate to make it something the UseStmt
itself controlled.  If we're using a module, return its block; if we're using
a TypeSymbol (for an EnumType, which should already be verified), return that
TypeSymbol as the scope.  Replace those implementations with calls to this
method.