Remove chpldoc-venv used to build the man pages in gen_release
With #3243 we now use rst2man.py from docutils to build the man pages. We use
to docutils from the chpldoc-venv, which means we build the chpldoc-venv to
create the man page.

However, we weren't removing the chpldoc-venv after building the man page,
which meant that the venv was getting bundled with the tarball, adding ~40 MB
uncompressed (~15-20 compressed)