Make use of 'use <enum>' in meteor
This was one of the few tests in release/examples that really felt
unwieldy to have to refer to an enum's name every time we wanted
to reference it.  As a result of Lydia's recent work, we now don't
need to.  As part of this change, I also made the enum have a more
descriptive name since I believe 'D' was chosen in part to avoid the
overhead of typing it every time...