Better handle code generation with declared extern C types
It has bothered me for some time that a program such as this:

   var x:c_uint;

would not generate C code that uses the type c_int (where c_int is some
extern type).

It also bothered me that the generated C code for SysCTypes.chpl had
several useless asserts. The Chapel code looks like this:

   assert(sizeof(c_uint) == sizeof(uint(32)));

where the intent is to fail early if there is something wrong with the
types generated in SysCTypes (e.g. we compile with c_ulong as 64-bits but
in fact it is 32-bits).

For context, note that types like c_uint are generated into
SysCTypes.chpl declarations that might look like this:

    extern type c_int= int(32);

This patch fixes both of these problems. These problems are slightly tied
together since they both have to do with how extern type type variables
work in the compiler.

== Keeping Declared Variable Extern Types ==

The basic problem is that the Chapel type system uses its version of
types instead of a declared type. That is a somewhat reasonable design
because then the compiler can normalize these types. However, the problem
comes up that for a variable like `x` in the example above, its type will
be `int(32)` rather than `c_int`.

This patch uses the approach of adding a field to VarSymbol that stores
the TypeSymbol for an explicitly declared type as in this case. This
connection is made in scopeResolve (which means that it will only work in
certain situations). Then, for VarSymbols with this TypeSymbol
associated, code generation uses the TypeSymbol instead of the type field
when generating the type of a variable.

== Cleaning Up sizeof(c_uint) ==

Initial versions of the above change caused problems with the
SysCTypes.chpl file. In particular, it turns out that something like
`sizeof(c_uint)` was code generating to `sizeof(tmp)` where `tmp` is a
local variable of type `uint(32)` in my case. That seemed obviously wrong
to me so I set about fixing it.

To fix it, I adjusted replaceTypeArgsWithFormalTypeTemps to pass a type
symbol to such calls instead of passing in a variable marked with
FLAG_TYPE_VARIABLE. In addition, I tightened up a workaround for code
like this

    extern var x: c_ptr(c_int)

that was replacing the a `c_uint` argument in a `sizeof` call with
`uint(32)` (for example). The workaround was to add code in
resolveDefaultGenericType in order to detect such cases and to replace
the argument with a new SymExpr for the TypeSymbol based on the
variable's type. This strategy had the drawback that it prevented these
sizeof calls in SysCTypes from working correctly. Instead, this patch
adjusts this workaround to call `resolveTypeAlias` and to store the
result in the VarSymbol's type. That strategy seems to fit slightly
better.