Add a pragma to declare an extern function safe for fast-on optimization
pragma "fast-on safe extern function" can be used to declare that an extern
function is safe for the fast-on optimization.  Any extern function without
this pragma will inhibit fast-on.

Added a test that declares extern functions with and without the pragma and
uses verboseComm to check if fast-on triggers for each.