Add test cases to cover three new Path functions (which are all related)
The functions in question are splitPath, basename, and dirname.  These tests
verify the behavior when passed an empty string, a path that is only the
path separator character for UNIX systems, a path that has no file portion,
a path that has no directory portion, a long path which ends in a file name,
a long path which ends in a directory, a short path with a directory and a file,
a path with a prefix of the path separator, a path with only the path separator
and a file, a path which contains multiple path separator characters in a row
followed by a file, and a file and directory that are separated by multiple
path separator characters.