Add implementation of basename, dirname, and splitPath functions to Path library
In reverse order:

splitPath takes a given path, which may or may not point to a valid location of
a file or directory, and returns a tuple pair of (directory name, file name)
as determined by the rightmost instance of the path separator character, if
present.  If the last character in the path is this separator character, then
the path in whole will be returned in the directory name portion.  If no
character in the path is this separator character, the entirety of the path
will be returned in the file name portion.  If multiple consecutive instances
of the path separator character occur between the directory name and the file
name, all will be removed.  The implementation for this function uses the
string method rfind to obtain the last instance of the path separator character,
then walks backwards until it has determined the correct end of the directory
name (based on how many consecutive path separator characters are present).

basename takes a given path and returns the file name portion of it, if present.
The implementation for this function calls splitPath and returns the second
field in the tuple obtained.

dirname takes a given path and returns the directory name portion of it, if
present.  The implementation for this functions calls splitPath and returns the
first field in the tuple obtained.