Optimize return of wide pointers
Many Chapel functions inside our internal modules are utility functions
that are frequently used to create or update objects.  Because these
functions are so commonly used, they often end up returning wide pointers.

One such example is the _array constructor. It looks a little like this in
the AST:

function _construct_array(_value : ArrayClass, meme : ArrayClass) {
  move this meme;
  move this value;
  call initialize this;
  return this;
}

Once this construct is called with wide arguments, it will always return
a wide pointer for every array in the program. This results in overhead
we would prefer to avoid.

As you can see, the return type is wide only because the arguments are
wide. But what if we have two different call sites like this?

  xx = _construct_array(narrow_temp, narrow_temp);

  yy = _construct_array(wide_temp, wide_temp);

If the compiler knew that _construct_array would return the same
wideness as its arguments, it could observe that `xx` should be a narrow
array. This commit adds this optimization for a couple of cases:

1) call_tmp = foo(arg)

2) _retArg = &call_tmp;
   foo(arg, _retArg);