Add some debug code to track jemalloc shared heap waste
Add some debug code to track how much memory is being used and how many
allocations are being made in order to waste memory that's not in the shared
heap.

Currently we just do a bunch of `sizeof(size_t)` allocations, but this can't be
very efficient. Preliminary runs suggest that we're doing over 130,000
allocations to waste just under a 1MB of memory (note that that's the
allocation size we request, jemalloc is probably rounding our allocations up to
be in some sort of size class)