Add code to check if all jemalloc size classes will be in the shared heap
Add some debug code (and helper routines) to determine if any small or large
size class will not be grabbed from our shared heap. Currently we waste memory
by just doing allocations of sizeof(size_t), but I wasn't sure if jemalloc
might have some fragments reserved for specific size chunks. It turns our that
it does, and it's possible for a few of the small size classes to not be
allocated out of the default memory jemalloc got from the system instead of our
shared heap.