Waste non-heap jemalloc memory faster (and smarter)
Previously we wasted memory with sizeof(size_t) allocations. Now we iterate
through the size classes from the largest large to the smallest small and
allocate until each size class is in the shared heap. The vastly reduces the
number of allocations (130,000 => 130) and means that we're no longer leaving
gaps for some allocation sizes to not come from the shared heap. Unfortunately
it also means that we're wasting ~14 MB (~12MB is part of the shared heap) in
order to do so.