Fix a sporadic failure to build jemalloc
We've been getting sporadic failures to build jemalloc with our OSX testng. We
got an error along the lines that the install directory already existed and it
couldn't be created again.

I believe what was happening is that the `install_lib_static`, `install_bin`,
and `install_include` targets were getting called simultaneously. Each one of
those targets was trying to create the install dir if it didn't exist, which is
racey.

To fix this I split out the install steps so that they are serialized.