Fix sizeof extern types e.g. sizeof(c_uint)
It has bothered me that the generated C code for SysCTypes.chpl had
several useless asserts. The Chapel code looks like this:

   assert(sizeof(c_uint) == sizeof(uint(32)));

where the intent is to fail early if there is something wrong with the
types generated in SysCTypes (e.g. we compile with c_ulong as 64-bits but
in fact it is 32-bits).

For context, note that types like c_uint are generated into
SysCTypes.chpl declarations that might look like this:

    extern type c_uint= uint(32);

`sizeof(c_uint)` was code generating to `sizeof(tmp)` where `tmp` is a local
variable of type `uint(32)` in my case. That seemed obviously wrong to me.
Then, the assert was comparing `sizeof(tmp)`, where `tmp` is a variable of type
`uint32_t`, against `sizeof(uint32_t)` - making the assert not function as
intended.

To fix it, I adjusted replaceTypeArgsWithFormalTypeTemps to pass a type symbol
to such calls instead of passing in a variable marked with FLAG_TYPE_VARIABLE.
In addition, I tightened up a workaround for code like this

    extern var x: c_ptr(c_int)

that was replacing the a `c_uint` argument in a `sizeof` call with `uint(32)`
(for example). The workaround was to add code in resolveDefaultGenericType in
order to detect such cases and to replace the argument with a new SymExpr for
the TypeSymbol based on the variable's type. This strategy had the drawback
that it prevented these sizeof calls in SysCTypes from working correctly. This
patch adjusts the workaround to call `resolveTypeAlias` and to store the result
in the VarSymbol's type. That strategy seems to fit slightly better.