Enable type aliases for generic class by removing chpl__typeAliasInit
A future work item from the addition of PCG to modules/standard/Random.chpl
is to adjust RandomStream to be NBPRandomStream and then (in some
future version) become PCGRandomStream. The obvious way to do that is
with a type alias e.g.:

  type RandomStream = PCGRandomStream;

However, since RandomStream is a generic class, this did not work as
expected.

This patch adjusts a few lines in normalize, resolution, and code
generation, in order to allow such type aliases to work. It removes
the chpl__typeAliasInit call which was previously added in normalize,
in favor of expecting function resolution to detect the error when
a type is initialized with a value.

For history, the chpl__typeAliasInit call was added in
68d1ad72860a37d7fbc36321fc00d4ad4316a7fd with the explanation "main
reason for creating and resolving this expression is to issue an error if
an attempt is made to initialize a type variable using a value that known
only at run-time."

Adds several tests that show the class type alias working in the case
needed for Random.chpl.

Also adds futures showing many related examples that do not work. I
believe that it would be very difficult to fix these examples with the
current handling of PRIM_NEW in normalize. I believe a fix would amount
to moving the "normalization" of PRIM_NEW to during resolution. Perhaps
this work can be considered along with initializer re-implementation.  (A
related improvement might be to store the constructor as a type method
instead of encoding the type-to-be-constructed in the constructor's name).