start_test should fail on logging errors
Override the error handling in Python's built-in logging module, to
make sure remotely-executed builds fail if something goes wrong with
the infrastructure (network connections, for example).
Normally, Python logging ignores such errors, on the assumption that
missing log messages are not very important to the application.
In this program, however, the logging provides an essential function.