Increase problem size for array-as-vec perf test
Bump the performance problem size from 30,000 to 50,000,000. This test has
gotten dramatically faster as a result of jemalloc and the recursive
doubling/halving. Increase the problem size so the kernels take more than a few
fractions of a second. There are some slower "linear" kernels (removeFront,
insert*) that take a lot longer, so I reduced the problem size of those
kernels.

On chap03 total execution time is ~72 second and on chapcs07 it's ~30 seconds
and most kernels take at least .5 seconds.

I also added perf annotations for this problem size increase and the
doubling/halving commit.