Report an error for a field or function in with-clause.
Syntactically, a with-clause consists of identifiers. Semantically,
those identifiers must refer to variables - global or local. If an
identifier in a with-clause instead references a field or paren-less
method of the current object or a paren-less function, we now generate
an error.

I put this check in this particular spot in scopeResolve. That's where
the compiler realizes that an UnresolvedSymExpr refers to a field or
function rather than a variable. So it's perfect to check whether it's
a with-clause, right there.

Alternatively we could iterate, separately, over all blocks or loops
that could have with clause. For each with clause, we could check
whether each element is a SymExpr or something else, and report an error
if it is the latter.