Add/fix annotations for this week
The array vector annotation for David's improvements needed to be applied more
broadly, as it affected more than just the test following it.  The other
annotation on that test was listed on the date the change went in, not the date
the change was observed (we've been trying to be consistent on that).  Also,
marked the improvement to memory leaks in examples from today's run to be Ben's
fixes to miniMD that went in yesterday.