Fix the date for the general array as vector change annotation
Since I took the annotation from the timeVectorArray section, the date it gave
was wrong (because the test didn't store timings the day after the change went
in due to a perfkeys mismatch).  Update the global annotation to give the
right date, and reinstate the annotation for timeVectorArray because annotations
only show up on days with data.